Fix: do not display graphiz when argument is set false
By default, addOptionalArgument discards an argument when the value
evaluates to false.
This results in the absence of the option in the final process command.

In many cases, the absence of an option means false, but in the case of
Deptrac's --formatter-graphviz-display it means true.

The use of addOptionalArgument will never allow false, so unsetting this
option in Deptrac is impossible.