Rework realpath logic to remove the possibility of a type error
We check that the file exists, but if for whatever reason we cant
make a realpath out of it, or if we have another process that deletes
this in the mean time, then we could have a type error.

By retrieving the real path, and then checking the error condition
we  remove the ability for a type error to occur here.

We dont change the fact that the file in question can still be
deleted in the mean time, but thats not within the scope of this method.